break up the style regex into 3 sub groups
Because I need to test that the beginning and ending number of quotes
are exactly the same I broke up the style regex into 3 groups for each
type.  Within each test I now also allow for a single quote to pass
through using the `[^']+'[^']+|[^']` syntax that looks for
non-quote+one-quote+non-quote text or all non-quote text.